refactor: move the rest of the integration stuff
