+++ src/components/ui/select.tsx
@@ -13,7 +13,7 @@ const SelectValue = React.forwardRef<
   React.ElementRef<typeof SelectPrimitive.Value>,
   React.ComponentPropsWithoutRef<typeof SelectPrimitive.Value>
 >(({ ...props }, ref) => {
-  return <SelectPrimitive.Value ref={ref} data-slot="select-value" className={cn('')} {...props} />
+  return <SelectPrimitive.Value ref={ref} data-slot="select-value" {...props} />
 })
 SelectValue.displayName = SelectPrimitive.Value.displayName
 
@@ -25,7 +25,7 @@ const SelectTrigger = React.forwardRef<
     ref={ref}
     data-slot="select"
     className={cn(
-      'relative flex w-full justify-between gap-2 rounded-2xl border border-light-10 bg-secondary px-3 pb-2 pt-7 text-[15px] text-primary outline-none transition-[padding] placeholder:text-subtle hover:border-light-20 disabled:cursor-not-allowed disabled:opacity-50 data-[state=open]:border-light-20 data-[state=open]:bg-tertiary [&>span]:line-clamp-1',
+      'relative flex w-full rounded-2xl border border-light-10 bg-secondary px-3 pb-2 pt-7 text-[15px] leading-snug text-primary outline-none transition-[padding] placeholder:text-subtle hover:border-light-20 disabled:cursor-not-allowed disabled:opacity-50 data-[state=open]:border-light-20 data-[state=open]:bg-tertiary [&>span]:line-clamp-1',
       className,
     )}
     value={value}
@@ -33,7 +33,7 @@ const SelectTrigger = React.forwardRef<
   >
     {children}
     <SelectPrimitive.Icon asChild>
-      <ChevronDownIcon className="h-4 w-4 opacity-50" />
+      <ChevronDownIcon className="absolute right-3 top-1/2 size-4 -translate-y-1/2 text-subtle" />
     </SelectPrimitive.Icon>
   </SelectPrimitive.Trigger>
 ))
