+++ src/features/integrations/types.ts
@@ -1,0 +1,4 @@
+/**
+ * All the types of integrations that we support.
+ */
+export type IntegrationType = 'linear' | 'notion'
